# docs/api/api_swagger.yaml
basePath: /api
definitions:
...
  github_com_slxar_rustdesk-api-goravel_v3_http_request_api.DeviceInfoInLogin:
    properties:
      name:
        type: string
      os:
        type: string
      type:
        type: string
    type: object
  github_com_slxar_rustdesk-api-goravel_v3_http_request_api.LoginForm:
    properties:
      autoLogin:
        type: boolean
      deviceInfo:
        $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_request_api.DeviceInfoInLogin'
      id:
        type: string
      password:
        maxLength: 32
        minLength: 4
        type: string
      type:
        type: string
      username:
        maxLength: 64
        minLength: 2
        type: string
      uuid:
        type: string
    required:
    - username
    type: object
  github_com_slxar_rustdesk-api-goravel_v3_http_request_api.PeerForm:
    properties:
      cpu:
        type: string
      hostname:
        type: string
      id:
        type: string
      memory:
        type: string
      os:
        type: string
      username:
        type: string
      uuid:
        type: string
      version:
        type: string
    type: object
...
  github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  github_com_slxar_rustdesk-api-goravel_v3_http_response_api.LoginRes:
    properties:
      access_token:
        type: string
      secret:
        type: string
      tfa_type:
        type: string
      type:
        type: string
      user:
        $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response_api.UserPayload'
    type: object
  github_com_slxar_rustdesk-api-goravel_v3_http_response_api.UserPayload:
    properties:
      avatar:
        type: string
      display_name:
        type: string
      email:
        type: string
      info:
        additionalProperties: true
        type: object
      is_admin:
        type: boolean
      name:
        type: string
      note:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: 接口
  title: 管理系统API
  version: "1.0"
paths:
...
  /currentUser:
    post:
      consumes:
      - application/json
      description: 用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response_api.UserPayload'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 用户信息
      tags:
      - 用户
...
  /login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 登录表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_request_api.LoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response_api.LoginRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      summary: 登录
      tags:
      - 登录
  /login-options:
    get:
      consumes:
      - application/json
      description: 登录选项
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      summary: 登录选项
      tags:
      - 登录
  /logout:
    post:
      consumes:
      - application/json
      description: 登出
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      summary: 登出
      tags:
      - 登录
...
  /sysinfo:
    post:
      consumes:
      - application/json
      description: 提交系统信息
      parameters:
      - description: 系统信息表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_request_api.PeerForm'
      produces:
      - application/json
      responses:
        "200":
          description: SYSINFO_UPDATED,ID_NOT_FOUND
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      summary: 提交系统信息
      tags:
      - System
  /sysinfo_ver:
    post:
      consumes:
      - application/json
      description: 获取系统版本信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      summary: 获取系统版本信息
      tags:
      - System
  /user/info:
    get:
      consumes:
      - application/json
      description: 用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response_api.UserPayload'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 用户信息
      tags:
      - 用户
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"